# openapi.yaml
openapi: 3.0.3
info:
  title: MediaHub API
  description: |
    Contrato mínimo gerado a partir de `server/server.mjs`.
    Autenticação: header `Authorization: Bearer <JWT>` quando indicado.
    Alguns endpoints de busca aceitam token opcional (chaves do provedor por usuário).
  version: "1.0.0"
servers:
  - url: http://localhost:8081
    description: API local (padrão `PORT` no .env)
  - url: https://api.example.com
    description: Substituir pela URL real em produção
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/password-reset/start:
    post:
      tags: [auth]
      summary: Inicia reset de senha
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
  /api/auth/password-reset/confirm:
    post:
      tags: [auth]
      summary: Confirma reset de senha
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
  /api/auth/password-reset/request:
    post:
      tags: [auth]
      summary: Solicita reset (fluxo alternativo)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
  /api/me:
    get:
      tags: [me]
      summary: Perfil autenticado
      responses:
        "200":
          description: Usuário
    put:
      tags: [me]
      summary: Atualiza perfil
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Atualizado
...
  /api/admin/password-reset-requests/{id}/reset:
    post:
      tags: [admin]
      summary: Executa reset por pedido
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK